create ItemNameHelper interface and it's implementation and use it in the ZipService

+++ src/main/java/app/Item.java
@@ -14,7 +14,7 @@ public class Item {
     public String url;
 
     @SerializedName("created_time")
-    public int createdTime;
+    public long createdTime;
 
     @SerializedName("type")
     public String type; // image or video
